+++ k210_layer.py
@@ -360,12 +360,66 @@ class K210Pool:
 
 
 class K210Layer:
-    def __init__(self, eight_bit_mode):
-        self.conv = None
-        self.bn = None
-        self.act = None
-        self.pool = None
-        self.eight_bit_mode = eight_bit_mode
+    def __init__(self, iwo_minmax, ico_shapes, conv_weights_isdw, bn_mean_var_gamma_beta_epsilon, act_type,
+                        pool_type_size_stride, eight_bit_mode=False, cbap_tensor_info=None, idx=-1):
+        input_min, input_max, weights_min, weights_max, output_min, output_max = iwo_minmax
+        input_shape, conv_shape, output_shape = ico_shapes
+        conv_weights, conv_isdw = conv_weights_isdw
+        conv_tensor_info, bn_tensor_info, act_tensor_info, pool_tensor_info, *_ = [
+            *list(cbap_tensor_info or []), dict(), dict(), dict(), dict()
+        ]
+
+        output_name = pool_tensor_info.get('name') or act_tensor_info.get('name', 'noname')
+        output_scale, output_bias = tools.min_max_to_scale_bias(output_min, output_max)
+        layer_shape_trans = [
+            int(input_shape[1]), int(input_shape[2]), int(input_shape[3]),
+            int(output_shape[1]), int(output_shape[2]), int(output_shape[3])
+        ]
+        print(
+            "[layer {}]".format(idx), output_name,
+            'shape(WHC): {}x{}x{} => {}x{}x{}'.format(*layer_shape_trans),
+            'scale/bias:', output_scale, output_bias
+        )
+
+        self.conv = K210Conv(
+            conv_weights,
+            conv_isdw,
+            eight_bit_mode, [input_shape, conv_shape],
+            [input_min, input_max, weights_min, weights_max],
+            tensor_info=conv_tensor_info
+        )
+
+        bn_mean, bn_var, bn_gamma, bn_beta, bn_epsilon = bn_mean_var_gamma_beta_epsilon
+        self.bn = K210BN(
+            bn_mean,
+            bn_var,
+            bn_gamma,
+            bn_beta,
+            bn_epsilon,
+            eight_bit_mode,
+            tensor_info=bn_tensor_info
+        )
+
+        self.act = K210Act(output_min, output_max, act_type,
+                                     eight_bit_mode=eight_bit_mode, tensor_info=act_tensor_info)
+
+        if pool_type_size_stride is not None:
+            pool_type, pool_size, pool_stride = pool_type_size_stride
+            if pool_size == 2 and conv_shape[3] % 2 != 0:
+                raise ValueError(
+                    "at {} unsupport padding mode SAME of pooling" \
+                        .format(pool_tensor_info.get('name', 'noname'))
+                )
+
+            if conv_isdw and pool_size != 1:
+                raise ValueError(
+                    'not supported DepthwiseConv2d({}) followed by pooling witch pool_size is not 1.' \
+                        .format(pool_tensor_info.get('name', 'noname'))
+                )
+
+            self.pool = K210Pool(pool_type, pool_size, pool_stride, pool_tensor_info)
+        else:
+            self.pool = None
 
     @staticmethod
     def batch(iter, n=1):
@@ -406,3 +460,100 @@ class K210Layer:
         dma_burst_size = 0xf
         send_data_out = 0
         return locals()
+
+
+
+def k210_layer_post_fix(kl_args_list):
+    def fix_dw_with_strde2(kl_args_list):
+        def expand_wh(shape_):
+            shape_1 = shape_[1] * 2
+            shape_2 = shape_[2] * 2
+            return [shape_[0], shape_1, shape_2, shape_[3]]
+
+        ret = []
+        lack_of_left_pooling = False
+        for kl_args in kl_args_list:
+            input_shape, conv_shape, output_shape = kl_args['ico_shapes']
+            conv_weights, conv_isdw = kl_args['conv_weights_isdw']
+            pool_type_size_stride = kl_args['pool_type_size_stride']
+            kl_args_fixed = dict(kl_args)
+
+            conv_kernel_size = int(conv_weights.shape[0])
+            conv_stride = int((int(input_shape[2])+1)/int(conv_shape[2]))
+
+            if lack_of_left_pooling:
+                if not conv_isdw and conv_kernel_size==1 and pool_type_size_stride is None:
+                    # fix in current layer
+                    input_shape = expand_wh(input_shape)
+                    conv_shape = expand_wh(conv_shape)
+                    lack_of_left_pooling = False
+                    kl_args_fixed['pool_type_size_stride'] = ['leftPool', 2, 2]
+                    kl_args_fixed['ico_shapes'] = [input_shape, conv_shape, output_shape]
+                else:
+                    if not (conv_kernel_size==1 and pool_type_size_stride is None):
+                        raise ValueError(
+                            'run fix_dw_with_strde2 failed. '+
+                            'can not delay left_pooling over current layer, '+
+                            'current layer conv_kernel_size:{}, pool_type_size_stride:{}'\
+                            .format(conv_kernel_size, pool_type_size_stride)
+                        )
+
+                    # delay fix in after layers
+                    input_shape = expand_wh(input_shape)
+                    conv_shape = expand_wh(conv_shape)
+                    output_shape = expand_wh(output_shape)
+                    kl_args_fixed['ico_shapes'] = [input_shape, conv_shape, output_shape]
+
+
+            if conv_stride == 2:
+                if not conv_isdw:
+                    if pool_type_size_stride is None:
+                        # fix in current layer
+                        conv_shape = expand_wh(conv_shape)
+                        kl_args_fixed['pool_type_size_stride'] = ['leftPool', 2, 2]
+                        kl_args_fixed['ico_shapes'] = [input_shape, conv_shape, output_shape]
+                    else:
+                        # fix later
+                        lack_of_left_pooling = True
+                        conv_shape = expand_wh(conv_shape)
+                        output_shape = expand_wh(output_shape)
+                        kl_args_fixed['ico_shapes'] = [input_shape, conv_shape, output_shape]
+                else:
+                    # dw layer needs to fix it later
+                    lack_of_left_pooling = True
+                    conv_shape = expand_wh(conv_shape)
+                    output_shape = expand_wh(output_shape)
+                    kl_args_fixed['ico_shapes'] = [input_shape, conv_shape, output_shape]
+
+
+            ret.append(kl_args_fixed)
+
+        if lack_of_left_pooling:
+            raise ValueError('run fix_dw_with_strde2 failed. no more layers for fix.')
+        return ret
+
+    def fix_wh_leas_than_4(kl_args_list):
+        def force_pad_to_4(shape_):
+                return [shape_[0], 4, 4, shape_[3]]
+
+        ret = []
+        for kl_args in kl_args_list:
+            input_shape, conv_shape, output_shape = kl_args['ico_shapes']
+            kl_args_fixed = dict(kl_args)
+
+
+            if input_shape[1] < 4 or conv_shape[1] < 4 or output_shape[1] < 4:
+                input_shape = force_pad_to_4(input_shape)
+                conv_shape = force_pad_to_4(conv_shape)
+                output_shape = force_pad_to_4(output_shape)
+                kl_args_fixed['ico_shapes'] = [input_shape, conv_shape, output_shape]
+
+            ret.append(kl_args_fixed)
+
+        return ret
+
+
+    kl_args_list = fix_wh_leas_than_4(kl_args_list)
+    kl_args_list = fix_dw_with_strde2(kl_args_list)
+    return kl_args_list
+
